generic: fix for wrong checksum calculation on 64bit systems - thx tr3b

+++ src/qcommon/md4.c
@@ -47,8 +47,8 @@ typedef unsigned char *POINTER;
 /* UINT2 defines a two byte word */
 typedef unsigned short int UINT2;
 
-/* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+/* UINT4 defines a four byte word - was unsigned long int which did not work for 64 bit */
+typedef unsigned int UINT4;
 
 
 /* MD4.H - header file for MD4C.C */
